requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt
# conda install --file requirements.txt

numpy
matplotlib
scipy
scikit-learn
pandas
pillow
pyyaml
tensorflow>=1.10.0

# if faiss-cpu fails to install using default channels, use
# conda install -c pytorch faiss-cpu


